+++ src/client/java/com/github/theredbrain/playerattributescreen/mixin/client/gui/screen/ingame/InventoryScreenMixin.java
@@ -69,8 +69,8 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 
 	@Inject(method = "init", at = @At("TAIL"))
 	protected void playerattributescreen$init(CallbackInfo ci) {
-		ClientConfig.GeneralClientConfig generalClientConfig = PlayerAttributeScreenClient.clientConfig.generalClientConfig;
-		this.showAttributeScreen = generalClientConfig.show_attribute_screen_when_opening_inventory_screen && !PlayerAttributeScreen.serverConfig.generalServerConfig.disable_player_attribute_screen;
+		ClientConfig generalClientConfig = PlayerAttributeScreenClient.CLIENT_CONFIG;
+		this.showAttributeScreen = generalClientConfig.show_attribute_screen_when_opening_inventory_screen.get() && !PlayerAttributeScreen.SERVER_CONFIG.disable_player_attribute_screen.get();
 		this.toggleShowAttributeScreenButton = this.addDrawableChild(new ToggleInventoryScreenWidget(this.x + 146, this.y + 7, this.showAttributeScreen, true, button -> {
 			if (!this.recipeBook.isOpen()) {
 				this.playerattributescreen$toggleShowAttributeScreen();
@@ -151,7 +151,7 @@ public abstract class InventoryScreenMixin extends AbstractInventoryScreen<Playe
 
 	@Unique
 	private void playerattributescreen$calculateRenderState() {
-		if (this.recipeBook.isOpen() || PlayerAttributeScreen.serverConfig.generalServerConfig.disable_player_attribute_screen) {
+		if (this.recipeBook.isOpen() || PlayerAttributeScreen.SERVER_CONFIG.disable_player_attribute_screen.get()) {
 			if (this.showAttributeScreen) {
 				this.playerattributescreen$toggleShowAttributeScreen();
 			}