+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0
+
+- fixed status effects not showing if the attribute screen is not visible
+- removed dependency on Cloth Config
+- added dependency on Fzzy Config
+
 # 1.0.0
 
 First release!